requirements.txt
requests 
pandas 
python-dotenv 
pandasql
langchain 
langchain-community 
langchain-openai
gradio
pdfplumber
matplotlib
seaborn
plotly
